Hopefully final iteration of ELFHandler as well as some cleaning
ELFHandler uses a java library which has also been included in the project as a .jar file. Since the library is neither hosted in a central repo (pip install or like not possible, import not possible) and were not using maven I see no other option than to include the .jar file.

Any tips/tricks that invalidate this reality is welcome.