Bugfixing ECALL with tests
Tests use ECALL to exit - not how it is configured in the processor (relinquishes control to exception handler located at MTVEC). None of these tests touch MTVEC, so it is always 0. We hence exit a test if we ever would jump to pc = 0. Not a very sustainable fix - should be edited in the future.