Added tests for hardware timer/counter/timer_events/edgecases
These tests required changing the frequency of the main clock (counters need to run at 100Hz) - as such I've added a frequency parameter to our classes and passed this down the pipeline. Default value is 100MHz so should work the implementation without any changes to other tests, programs, or code.